backend/requirements.txt
fastapi
uvicorn
google-genai
genanki
python-multipart
pydantic
PyPDF2
tenacity
